Moved files into separate directories for generation code update
