+++ admin/src/resources/AdminNote.js
